refactor: 로그인, 로그아웃 분리
