+++ src/main/java/kr/codesqaud/cafe/config/SignInInterceptor.java
@@ -31,6 +31,6 @@ public class SignInInterceptor implements HandlerInterceptor {
             throw new ApiUnauthorizedException();
         }
 
-        response.sendRedirect("/members/sign-in");
+        response.sendRedirect("/sign-in");
     }
 }
