+++ src/main/java/kr/codesqaud/cafe/exception/common/BadRequestExceptionView.java
@@ -2,12 +2,12 @@ package kr.codesqaud.cafe.exception.common;
 
 import java.util.Arrays;
 import kr.codesqaud.cafe.dto.member.ProfileEditRequest;
-import kr.codesqaud.cafe.dto.member.SignInRequest;
+import kr.codesqaud.cafe.dto.authentication.SignInRequest;
 import kr.codesqaud.cafe.dto.member.SignUpRequest;
 
 public enum BadRequestExceptionView {
 
-    SIGN_IN_FORM(SignInRequest.class, "member/signIn"),
+    SIGN_IN_FORM(SignInRequest.class, "authentication/signIn"),
     SIGN_UP_FORM(SignUpRequest.class, "member/signUp"),
     PROFILE_EDIT_FORM(ProfileEditRequest.class, "member/profileEdit"),
     EMPTY(Object.class, "error/500");
