+++ src/main/java/kr/codesqaud/cafe/service/MemberService.java
@@ -5,7 +5,6 @@ import java.util.stream.Collectors;
 import kr.codesqaud.cafe.domain.Member;
 import kr.codesqaud.cafe.dto.member.MemberResponse;
 import kr.codesqaud.cafe.dto.member.ProfileEditRequest;
-import kr.codesqaud.cafe.dto.member.SignInRequest;
 import kr.codesqaud.cafe.dto.member.SignUpRequest;
 import kr.codesqaud.cafe.exception.common.UnauthorizedException;
 import kr.codesqaud.cafe.exception.member.MemberDuplicateEmailException;
@@ -33,20 +32,6 @@ public class MemberService {
     private void validateDuplicateEmail(SignUpRequest signUpRequest) {
         if (memberRepository.existsByEmail(signUpRequest.getEmail())) {
             throw new MemberDuplicateEmailException(signUpRequest);
-        }
-    }
-
-    @Transactional(readOnly = true)
-    public MemberResponse signIn(SignInRequest signInRequest) {
-        Member member = memberRepository.findByEmail(signInRequest.getEmail())
-            .orElseThrow(() -> new MemberInvalidPassword(signInRequest));
-        validateSamePassword(member, signInRequest);
-        return MemberResponse.from(member);
-    }
-
-    private void validateSamePassword(Member member, SignInRequest signInRequest) {
-        if (!member.equalsPassword(signInRequest.getPassword())) {
-            throw new MemberInvalidPassword(signInRequest);
         }
     }
