+++ src/main/resources/errorMessages.properties
@@ -1,6 +1,6 @@
 NotBlank.email= 이메일은 필수입니다.
 NotBlank.password= 비밀번호는 필수입니다.
-NotBlank.nickName= 닉네임은 필수입니다.
+NotBlank.nickname= 닉네임은 필수입니다.
 NotBlank.title= 제목은 필수입니다.
 NotBlank.content= 내용은 필수입니다.
 
@@ -8,8 +8,8 @@ Email.email= 이메일 형식에 맞게 입력해주세요.
 
 Pattern.password= 알파벳 대문자, 소문자 숫자가 각각 1개 이상이며, 길이가 8자리 이상 32글자 이하여야 합니다.
 
-Length.signUpRequest.nickName= 최소 {2}글자, 최대 {1}글자입니다.
-Length.profileEditRequest.nickName= 최소 {2}글자, 최대 {1}글자입니다.
+Length.signUpRequest.nickname= 최소 {2}글자, 최대 {1}글자입니다.
+Length.profileEditRequest.nickname= 최소 {2}글자, 최대 {1}글자입니다.
 Length.postWriteRequest.title= 최소 {2}글자, 최대 {1}글자입니다.
 Length.postWriteRequest.content= 최소 {2}글자, 최대 {1}글자입니다.
