+++ src/main/resources/templates/layout/header.html
@@ -5,10 +5,10 @@
         </a>
         <div id="endHeader" class="col-3  text-end">
             <th:block th:if="${session.account == null}">
-                <a th:href="@{/members/sign-in}" class="btn btn-primary btn-sm">로그인</a>
-                <a th:href="@{/members/sign-up}" class="btn btn-primary btn-sm">회원가입</a>
+                <a th:href="@{/sign-in}" class="btn btn-primary btn-sm">로그인</a>
+                <a th:href="@{members/sign-up}" class="btn btn-primary btn-sm">회원가입</a>
             </th:block>
-            <form th:if="${session.account != null}" th:action="@{/members/sign-out}" th:method="post">
+            <form th:if="${session.account != null}" th:action="@{/sign-out}" th:method="post">
                 <a th:href="@{/members}" class="btn">멤버 리스트</a>
                 <button class="btn btn-primary btn-sm" type="submit">로그아웃</button>
                 <a th:href="@{/members/{id}/form(id = ${session.account.id})}" class="btn btn-primary btn-sm">개인정보수정</a>
